Refactor to extract spider
